x11drv: rawinput should use unaccelerated output
Without unacclerated mouse input, gaming with a mouse is a really bad
experience. Changing mouse pointer speed for the desktop also changes it
for the game. It is really important for competitive first person
shooter games.

This is a major rewrite of xinput2 handling. It depends on the
wine-staging patches that add the SEND_HWMSG_RAWINPUT flag.

The new logic is this: whenever a window is focused, subscribe to
xinput2 events. When a window is unfocused, unsubscribe from xinput2
events.

If there are any unfocused windows that want rawinput, they need to set
RIDEV_INPUTSINK and wineserver will be responsible for delivering events
to them.

I bumped the XInput2 version request to 2.1 (released Dec 2011) to make
it clear that grabbing is not relevant for raw input. To
quote from xorg-xserver:
  XI 2.0: events delivered to the grabbing client (if any) OR to all root windows
  XI 2.1: events delivered to all root windows, regardless of grabbing state.